+++ wand/api.py
@@ -249,7 +249,8 @@ try:
 
     library.MagickWriteImageFile.argtypes = [ctypes.c_void_p, ctypes.c_void_p]
 
-    library.MagickWriteImages.argtypes = [ctypes.c_void_p, ctypes.c_char_p]
+    library.MagickWriteImages.argtypes = [ctypes.c_void_p, ctypes.c_char_p,
+                                          ctypes.c_int]
 
     library.MagickWriteImagesFile.argtypes = [ctypes.c_void_p, ctypes.c_void_p]
 
@@ -271,6 +272,11 @@ try:
 
     library.MagickGetImageHeight.argtypes = [ctypes.c_void_p]
     library.MagickGetImageHeight.restype = ctypes.c_size_t
+
+    library.MagickGetImageOrientation.argtypes = [ctypes.c_void_p]
+    library.MagickGetImageOrientation.restype = ctypes.c_int
+
+    library.MagickSetImageOrientation.argtypes = [ctypes.c_void_p, ctypes.c_int]
 
     library.MagickGetImageUnits.argtypes = [ctypes.c_void_p]